fix(libs/NEO_6M): mejora la robustez y precisión de la librería GPS
*   Protege el acceso a `gpsData` con deshabilitación/habilitación de IRQ para garantizar la integridad de los datos.
*   Declara `rxData` como `volatile` para asegurar un manejo correcto en interrupciones.
*   Previene desbordamientos de búfer utilizando `snprintf` con límite y validación del tamaño de `rxBuffer`.
*   Asegura la correcta desinicialización de la recepción UART con `HAL_UART_AbortReceive_IT`.
*   Ajusta la declaración de `latitude` y `longitude` a `double` en el header, completando la implementación de mayor precisión.
*   Corrige advertencias de tipo en llamadas a `HAL_UART_Receive_IT` y añade inclusiones de headers necesarias.

+++ Librerias/NEO_6M/NEO_6M.h
@@ -22,6 +22,9 @@
 
 #include "main.h"
 #include <stdint.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -53,8 +56,8 @@ typedef enum {
  * @brief Estructura de datos que almacena el último fix GPS parseado.
  */
 typedef struct {
-    float latitude;     /**< Latitud en grados decimales          */
-    float longitude;    /**< Longitud en grados decimales         */
+    double latitude;    /**< Latitud en grados decimales          */
+    double longitude;   /**< Longitud en grados decimales         */
     char  northSouth;   /**< Indicador de hemisferio: 'N' o 'S'   */
     char  eastWest;     /**< Indicador de hemisferio: 'E' o 'W'   */
     uint8_t  hours;     /**< Hora UTC del fix (GPGGA/GPRMC)       */
